streamlit/requirements.txt
streamlit>=1.28.0
tencentcloud-sdk-python>=3.0.1200
